requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy[asyncio]==2.0.30
asyncmy==0.2.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.2.1
httpx==0.27.0
ruff==0.4.4
pytest-cov==5.0.0
loguru==0.7.2
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
